refactor(websocket): seperate modules
